app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asyncpg==0.30.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
blinker==1.9.0
boto3==1.26.137
botocore==1.29.165
celery==5.3.4
certifi==2025.6.15
cffi==1.17.1
channels==4.3.0
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==2.3.0
Django==5.1.11
django-allauth==0.57.0
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-otp==1.5.4
django-ratelimit==4.1.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
geographiclib==2.0
geopy==2.3.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.2
msgpack==1.1.1
mysql-connector-python==9.4.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.5.0
python3-openid==3.2.0
qrcode==8.0
redis==5.0.1
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.6.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
virtualenv==20.32.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.interface==7.2
